requirements.txt
requests
pytest
pyyaml
faker
allure-pytest